Increase maxReplicas to 150
As per the BF Investigation of Scalings, I found out that a safer maxReplicas for this app would be 150.

Here's the full list:
https://github.com/vtex/io-infra-tools/blob/master/bf-2020/critical-apps/scalings.txt#L27
On the list I had 120 for this app, but I saw now that the regular usage
is already pretty high (30 replicas), so I decided to increase it a bit
more to be more future-proof.