encodeURI on pageType calls to encode unicode characters
Cyrillic characters were being silently dropped when the pageType
API was called by searchMetadata, this caused data to not be
returned by the resolver.

This was fixed by adding a encodeURI on the search client call to
the pageType API and also start to log all failures to call this
API on call site.